feat: add new actionable tile component + rename folder
