+++ src/ui/components/common/actionables/ActionableTile.css
@@ -1,0 +1,17 @@
+.actionableTile {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0 0.5em;
+    background-color: var(--mantine-color-gray-0, #f8f9fa);
+    border: 1px solid var(--mantine-color-gray-3, #dee2e6);
+    border-radius: 6px;
+}
+
+.actionableTile:hover {
+    background-color: rgba(0, 0, 0, 0.08);
+}
+
+.actionableTile .mantine-ActionIcon-root {
+    background-color: transparent !important;
+}
