+++ src/ui/components/common/actionables/actions/CloseActionIcon.tsx
