+++ src/ui/components/common/actionables/actions/CopyActionIcon.tsx
