+++ src/ui/components/common/actionables/actions/DeleteActionIcon.tsx
