+++ src/ui/components/common/actionables/actions/EditActionIcon.tsx
