+++ src/ui/components/common/actionables/actions/RevertActionIcon.tsx
